+++ src/net/xiaoluo/crazyit/crazyjava/annotations/RunTests.java
@@ -1,0 +1,8 @@
+package net.xiaoluo.crazyit.crazyjava.annotations;
+
+public class RunTests {
+    public static void main(String[] args) throws Exception {
+        ProcessorTest.process("net.xiaoluo.crazyit.crazyjava.annotations.MyTest");
+//        MyTest mt = new MyTest();
+    }
+}