+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Crazy Java Practice Project #
 
+## 2018-12-04 ##
+### Added ###
+- Chapter 14 "Annotations" example codes, some simple codes are excluded, 
+as well as the final processor code(need more understanding on that)
+
+## 2018-12-03 ##
+### Added ###
+- Chapter 10 "Exceptions" example codes, some simple codes are excluded
+
 ## 2018-11-29 ##
 ### Added ###
 - Chapter 7 "Fundamental Java Libs" example codes, some simple codes are excluded